core-assets/tessellating-pbr-generator/requirements.txt
aiofiles>=23.2.1
aiohttp>=3.9.0
jsonschema>=4.19.0
numpy>=1.24.0
opencv-python>=4.8.0
openai>=1.0.0
Pillow>=10.0.0
python-dotenv>=1.0.0
requests>=2.31.0
scipy>=1.11.0
tqdm>=4.66.0
typing-extensions>=4.8.0